chore: migrate grpc to instrumentation class to avoid expose internals
