--- packages/opentelemetry-instrumentation-grpc/src/grpc/client/patches.ts
@@ -1,45 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type * as grpcTypes from 'grpc';
-import { GrpcInstrumentation } from '../../instrumentation';
-import { getMethodsToWrap, getPatchedClientMethods } from './utils';
-
-export const patchClient = function (
-  this: GrpcInstrumentation,
-  grpcClient: typeof grpcTypes
-) {
-  const instrumentation = this;
-  return (original: typeof grpcTypes.makeGenericClientConstructor): never => {
-    instrumentation.getLogger().debug('patching client');
-    return function makeClientConstructor(
-      this: typeof grpcTypes.Client,
-      methods: { [key: string]: { originalName?: string } },
-      _serviceName: string,
-      _options: grpcTypes.GenericClientOptions
-    ) {
-      const client = original.apply(this, arguments as any);
-      instrumentation
-        .getShimmer()
-        .massWrap(
-          client.prototype as never,
-          getMethodsToWrap(instrumentation, client, methods) as never[],
-          getPatchedClientMethods.call(instrumentation, grpcClient) as any
-        );
-      return client;
-    } as never;
-  };
-};
